Fixed two tutorial bugs.
Official tutorial says in a paragraph:

"The resulting image file ashello5.img has 248 bytes", but then an "ls -l" is offered to confirm and it correctly shows that the size is 132. It's gonna be 248 when the file is assembled with all .asm providing referenced symbols.

The other fix is a bug in ashello.asm. It was missing a fundamental "DEF START" at the preamble. The omission was causing the symbol not to be included in the .obj. Consequently, when entering START as program name in E/A #3 (Load and Run), it was causing a "PROGRAM NOT FOUND" error.