Fixed errors in ashello_new.asm
The new syntax source file had a non-honored reference to kscan (non-honored: the corresponding copy was missing).
I fixed it by copying "keyscan_ea.asm", although for that one to work, I needed to replace the "bl kscan" (branch-and-link) command with "blwp kscan" (branch-and-load-workspace-pointer).

Not copying the file had a consequence: the program image (option -i) couldn't be created unless kscan_ea.asm was added to the command line every time. But being other copies already there, why not adding this one so this example is a stand-alone case?
Not changing "bl" in favor of "blwp" had another consequence: kscan was actually messing things up and the screen border wasn't really changing in response to pressed keys.